added organization to the app
