+++ client/package.json
@@ -9,6 +9,7 @@
     "cra-template": "1.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.2.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^4.2.4"
   },
